Make adding buffers more like adding parameters
Add similar semantics for creating a buffer object similar to creating
a Parameter. This is done by introducing a new Buffer class that can
be used for type disambiguation. The underlying functionality of
registering a buffer remains the same as the register_buffer method
has not been changed. A persistent parameter is present in Buffer type
to indicate whether a buffer should be persistent or not. Other non-test
changes have to do with getting the new Buffer type recognized by
inductor and dynamo. Remaining changes are test changes to make sure that
the Buffer type can be used as a drop in replacement for register_buffer
as it just leads to register_buffer being called.