Add timeout for translation validation instances.
As of now, translation validation runs to its completion. However, Z3 is time
consuming. PR #104464, for example, disables translation validation for a few benchmarks.

Instead, this PR introduces a timeout for translation validation. In that case, Z3 will
return `unknown`, since it wasn't able to prove or disprove the assertions. Then, we log
it as a warning, but don't stop execution.

Here's a summary of the changes:

- Added an environment variable for turning translation validation on and off
- Added an environment variable for setting the translation validation timeout
- Possibly reverts the changes in #104464
- Move from "QF_NRA" to "QF_NIRA" logic
    - It makes more sense, given the nature of the problems

ghstack-source-id: 87955a83d8d2a76800ca49ac47d6584f55b0d154
Pull Request resolved: https://github.com/pytorch/pytorch/pull/104654